Refactor: Make stats tab reusable, implement it as user's stats tab too
