+++ src/pages/Wallets/WalletDetailTable/WalletDetailTable.module.css
